code refactoring; parsing all audio headers
